requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2
fonttools==4.39.3
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
psycopg2-binary==2.9.9
six==1.16.0
sqlparse==0.4.4
tk==0.1.0
zipp==3.15.0
djangorestframework==3.14.0
django-cors-headers==4.3.0